remove WalletAdvancedProvider
