--- src/wallet/components/WalletAdvancedProvider.tsx
@@ -1,94 +1,0 @@
-import { RequestContext } from '@/core/network/constants';
-import { useValue } from '@/internal/hooks/useValue';
-import { usePortfolio } from '@/wallet/hooks/usePortfolio';
-import { type ReactNode, createContext, useContext, useState } from 'react';
-import type {
-  WalletAdvancedContextType,
-  WalletAdvancedFeature,
-} from '../types';
-import { useWalletContext } from './WalletProvider';
-
-type WalletAdvancedProviderReact = {
-  children: ReactNode;
-};
-
-const emptyContext = {} as WalletAdvancedContextType;
-
-const WalletAdvancedContext =
-  createContext<WalletAdvancedContextType>(emptyContext);
-
-export function useWalletAdvancedContext() {
-  const walletAdvancedContext = useContext(WalletAdvancedContext);
-  if (walletAdvancedContext === emptyContext) {
-    throw new Error(
-      'useWalletAdvancedContext must be used within a WalletAdvancedProvider',
-    );
-  }
-  return walletAdvancedContext;
-}
-
-export function WalletAdvancedProvider({
-  children,
-}: WalletAdvancedProviderReact) {
-  const { address, isSubComponentClosing, showSubComponentAbove } =
-    useWalletContext();
-  const [activeFeature, setActiveFeature] =
-    useState<WalletAdvancedFeature | null>(null);
-  const [isActiveFeatureClosing, setIsActiveFeatureClosing] = useState(false);
-  const {
-    data: portfolioData,
-    refetch: refetchPortfolioData,
-    isFetching: isFetchingPortfolioData,
-    dataUpdatedAt: portfolioDataUpdatedAt,
-  } = usePortfolio({ address }, RequestContext.Wallet);
-
-  const portfolioFiatValue = portfolioData?.portfolioBalanceInUsd;
-  const tokenBalances = portfolioData?.tokenBalances;
-
-  const animations = getAnimations(
-    isSubComponentClosing,
-    showSubComponentAbove,
-  );
-
-  const value = useValue({
-    activeFeature,
-    setActiveFeature,
-    isActiveFeatureClosing,
-    setIsActiveFeatureClosing,
-    tokenBalances,
-    portfolioFiatValue,
-    isFetchingPortfolioData,
-    portfolioDataUpdatedAt,
-    refetchPortfolioData,
-    animations,
-  });
-
-  return (
-    <WalletAdvancedContext.Provider value={value}>
-      {children}
-    </WalletAdvancedContext.Provider>
-  );
-}
-
-function getAnimations(
-  isSubComponentClosing: boolean,
-  showSubComponentAbove: boolean,
-) {
-  if (isSubComponentClosing) {
-    return {
-      container: showSubComponentAbove
-        ? 'fade-out slide-out-to-bottom-1.5 animate-out fill-mode-forwards ease-in-out'
-        : 'fade-out slide-out-to-top-1.5 animate-out fill-mode-forwards ease-in-out',
-      content: '',
-    };
-  }
-
-  return {
-    container: showSubComponentAbove
-      ? 'fade-in slide-in-from-bottom-1.5 animate-in duration-300 ease-out'
-      : 'fade-in slide-in-from-top-1.5 animate-in duration-300 ease-out',
-    content: showSubComponentAbove
-      ? 'fade-in slide-in-from-bottom-2.5 animate-in fill-mode-forwards duration-300 ease-out'
-      : 'fade-in slide-in-from-top-2.5 animate-in fill-mode-forwards duration-300 ease-out',
-  };
-}
