+++ src/wallet/components/WalletProvider.tsx
@@ -13,8 +13,10 @@ import {
 } from 'react';
 import type { ReactNode } from 'react';
 import { useAccount } from 'wagmi';
-import type { WalletContextType } from '../types';
+import type { WalletAdvancedFeature, WalletContextType } from '../types';
 import { calculateSubComponentPosition } from '../utils/getWalletSubComponentPosition';
+import { RequestContext } from '@/core/network/constants';
+import { usePortfolio } from '../hooks/usePortfolio';
 
 const emptyContext = {} as WalletContextType;
 
@@ -34,6 +36,24 @@ export function WalletProvider({ children }: WalletProviderReact) {
   const connectRef = useRef<HTMLDivElement>(null);
   const { address } = useAccount();
   const breakpoint = useBreakpoints();
+
+  const [activeFeature, setActiveFeature] =
+    useState<WalletAdvancedFeature | null>(null);
+  const [isActiveFeatureClosing, setIsActiveFeatureClosing] = useState(false);
+  const {
+    data: portfolioData,
+    refetch: refetchPortfolioData,
+    isFetching: isFetchingPortfolioData,
+    dataUpdatedAt: portfolioDataUpdatedAt,
+  } = usePortfolio({ address }, RequestContext.Wallet);
+
+  const portfolioFiatValue = portfolioData?.portfolioBalanceInUsd;
+  const tokenBalances = portfolioData?.tokenBalances;
+
+  const animations = getAnimations(
+    isSubComponentClosing,
+    showSubComponentAbove,
+  );
 
   const handleClose = useCallback(() => {
     if (!isSubComponentOpen) {
@@ -65,6 +85,16 @@ export function WalletProvider({ children }: WalletProviderReact) {
     connectRef,
     showSubComponentAbove,
     alignSubComponentRight,
+    activeFeature,
+    setActiveFeature,
+    isActiveFeatureClosing,
+    setIsActiveFeatureClosing,
+    tokenBalances,
+    portfolioFiatValue,
+    isFetchingPortfolioData,
+    portfolioDataUpdatedAt,
+    refetchPortfolioData,
+    animations,
   });
 
   return (
@@ -74,4 +104,27 @@ export function WalletProvider({ children }: WalletProviderReact) {
 
 export function useWalletContext() {
   return useContext(WalletContext);
+}
+
+function getAnimations(
+  isSubComponentClosing: boolean,
+  showSubComponentAbove: boolean,
+) {
+  if (isSubComponentClosing) {
+    return {
+      container: showSubComponentAbove
+        ? 'fade-out slide-out-to-bottom-1.5 animate-out fill-mode-forwards ease-in-out'
+        : 'fade-out slide-out-to-top-1.5 animate-out fill-mode-forwards ease-in-out',
+      content: '',
+    };
+  }
+
+  return {
+    container: showSubComponentAbove
+      ? 'fade-in slide-in-from-bottom-1.5 animate-in duration-300 ease-out'
+      : 'fade-in slide-in-from-top-1.5 animate-in duration-300 ease-out',
+    content: showSubComponentAbove
+      ? 'fade-in slide-in-from-bottom-2.5 animate-in fill-mode-forwards duration-300 ease-out'
+      : 'fade-in slide-in-from-top-2.5 animate-in fill-mode-forwards duration-300 ease-out',
+  };
 }
